+++ CompuZone/CompuZone.PL/Controllers/ProductController.cs
@@ -1,5 +1,6 @@
 ﻿using CompuZone.BLL.DTOs.Product;
 using CompuZone.BLL.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -21,8 +22,13 @@ namespace CompuZone.PL.Controllers
             return Ok(result); 
         }
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> Create([FromBody] ReqProductDto dto)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return Unauthorized("You are not an Admin."); // Returns 401
+            }
             var result = await _service.CreateAsync(dto);
             return Ok(result);
         }
@@ -33,14 +39,24 @@ namespace CompuZone.PL.Controllers
             return Ok(result);
         }
         [HttpPut("{id}")]
+        [Authorize]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] ReqProductDto dto)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return Unauthorized("You are not an Admin."); // Returns 401
+            }
             var result = await _service.UpdateAsync(id, dto);
             return Ok(result);
         }
         [HttpDelete("{id}")]
+        [Authorize]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!User.IsInRole("Admin"))
+            {
+                return Unauthorized("You are not an Admin."); // Returns 401
+            }
             var result = await _service.DeleteAsync(id);
             return Ok(result);
         }
